add Semaphore64
Create a simple implementation of 64-bit semaphore.
FlowController's byte-limit is migrated to use it.

setMaxOutstandingRequestBytes(Integer) method is added
to maintain source compatibility with google-cloud-java.
This is necessary since an int can be implicitly converted to
an Integer or a long but not a Long.

+++ gax/src/test/java/com/google/api/gax/batching/ThresholdBatcherTest.java
@@ -73,7 +73,7 @@ public class ThresholdBatcherTest {
   private static TrackedFlowController trackedFlowController;
 
   private static BatchingFlowController<SimpleBatch> getTrackedIntegerBatchingFlowController(
-      Integer elementCount, Integer byteCount, LimitExceededBehavior limitExceededBehaviour) {
+      Integer elementCount, Long byteCount, LimitExceededBehavior limitExceededBehaviour) {
     trackedFlowController =
         new TrackedFlowController(
             FlowControlSettings.newBuilder()
